+++ test/client/bulk_test.rb
@@ -293,4 +293,31 @@ describe Elastomer::Client::Bulk do
     assert_equal 'tweet', items[0]['index']['_type']
     assert_equal 1, items[0]['index']['_version']
   end
+
+  it 'streams bulk responses' do
+    ops = [
+      [:index, { :message => 'tweet 1' }, { :_id => 1, :_type => 'book', :_index => @index.name }],
+      [:index, { :message => 'tweet 2' }, { :_id => 2, :_type => 'book', :_index => @index.name }],
+      [:index, { :message => 'tweet 3' }, { :_id => 3, :_type => 'book', :_index => @index.name }]
+    ]
+    responses = $client.bulk_stream_responses(ops, { :action_count => 2 }).to_a
+    assert_equal(2, responses.length)
+    assert_bulk_index(responses[0]["items"][0])
+    assert_bulk_index(responses[0]["items"][1])
+    assert_bulk_index(responses[1]["items"][0])
+  end
+
+  it 'streams bulk items' do
+    ops = [
+      [:index, { :message => 'tweet 1' }, { :_id => 1, :_type => 'book', :_index => @index.name }],
+      [:index, { :message => 'tweet 2' }, { :_id => 2, :_type => 'book', :_index => @index.name }],
+      [:index, { :message => 'tweet 3' }, { :_id => 3, :_type => 'book', :_index => @index.name }]
+    ]
+    items = []
+    stats = $client.bulk_stream_items(ops, { :action_count => 2 }) { |item| items << item }
+    assert_equal(3, items.length)
+    assert_bulk_index(items[0])
+    assert_bulk_index(items[1])
+    assert_bulk_index(items[2])
+  end
 end